requirements.txt
asyncpg
psycopg
sqlalchemy

fastapi
pydantic
pydantic_settings
uvicorn
pyjwt
python-multipart